requirements.txt
streamlit
stripe
streamlit-extras
st-supabase-connection
streamlit-elements